requirements.txt
beautifulsoup4
langchain
groq
langchain-groq
langchain-community
python-dotenv
pypdf
Flask
google-search-results
chromadb